requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
soupsieve==2.5
yarl==1.9.4
